WIP * Start of scripting redesign * Initial version of VERM->Lua converter * Added basic test for Timed spell effect * Added a test for exiting 2 ERM examples (no NetPacks checked yet) * started spell effect API tests * started ScriptHandler * Added basic script context caching * Started Lua script module * Started Lua spell effect API * CMake/CI: fix compilation with LuaJIT on Linux * CI: fix building with LuaJIT on Mac * Started script state persistence * Started battle info callback binding * Added !!BU:C, !!BU:E, !!BU:G, !!BU:M * Started populating include directory with handlers abstractions * First working example with events. * Added !!MF:N + some refactoring * Get rid of CommitPackage package * C::B: added more test data to project * Use only BattleLogMessage for battle log transferring. * Extracted spells::Caster to own header; Expanded Spell API. * Started work on !!BM receiver

+++ client/Graphics.cpp
@@ -428,7 +428,7 @@ void Graphics::addImageListEntry(size_t index, std::string listName, std::string
 	if (!imageName.empty())
 	{
 		JsonNode entry;
-		entry["frame"].Float() = index;
+		entry["frame"].Integer() = index;
 		entry["file"].String() = imageName;
 
 		imageLists["SPRITES/" + listName]["images"].Vector().push_back(entry);
